backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Task Manager backend with FastAPI"
authors = ["wyattp23 <wyattp23@gmail.com>"]
readme = "README.md"

package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.115.5"}
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^4.2.1"
pytest = "^8.3.3"
httpx = "^0.27.2"
argon2-cffi = "^23.1.0"
pytest-asyncio = "^0.24.0"
psycopg2-binary = "^2.9.9"
psycopg2 = "^2.9.10"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
httpx = "^0.27.2"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "strict"
